remove merkle tree merge/par inserts for now
