+++ src/tests/mod.rs
@@ -1,6 +1,7 @@
 mod madara_comparison;
 mod merge;
 mod proof;
+mod proptest;
 mod simple;
 mod transactional_state;
 mod trie_log;
