+++ src/trie/mod.rs
@@ -2,6 +2,5 @@ mod merkle_node;
 pub mod merkle_tree;
 mod path;
 mod trie_db;
-mod merge;
 
 pub(crate) use trie_db::TrieKey;
